fixing compilers warnings

+++ Aiko/src/opengl/rawModel.cpp
@@ -3,7 +3,7 @@
 namespace opengl
 {
 
-    RawModel::RawModel(GLuint vao, unsigned int vertexCount)
+    RawModel::RawModel(GLuint vao, size_t vertexCount)
     {
         m_vaoID = vao;
         m_vertexCount = vertexCount;
